fix : cart logic

+++ src/controller/implementation/cart/cart-controller.ts
@@ -9,7 +9,8 @@ export class CartController {
   constructor(private cartservice: CartService) {}
   async addItemToCart(req: AuthRequest, res: Response) {
     try {
-      if(req.user){
+      
+      if (req.user) {
         const newCart = req.body;
         const { message, success } = await this.cartservice.addToCart(
           req.user.id,
@@ -17,11 +18,13 @@ export class CartController {
         );
         if (success) res.status(HttpStatus.CREATED).json({ message, success });
         else res.status(HttpStatus.BAD_REQUEST).json({ message, success });
-      } else  res.status(HttpStatus.UNAUTHORIZED).json({ 
-        message: responseMessage.TOKEN_ACCESS, 
-        success: false 
-      });
+      } else
+        res.status(HttpStatus.UNAUTHORIZED).json({
+          message: responseMessage.TOKEN_ACCESS,
+          success: false,
+        });
     } catch (error) {
+      console.log((error as Error).message);
       res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
         response: responseMessage.ERROR_MESSAGE,
         error: (error as Error).message,
@@ -31,14 +34,20 @@ export class CartController {
 
   async updateCart(req: AuthRequest, res: Response) {
     try {
-      const cartId = req.params.id as string;
-      const updateData = req.body;
-      const { message, success } = await this.cartservice.updateCart(
-        cartId,
-        updateData
-      );
-      if (success) res.status(HttpStatus.OK).json({ message, success });
-      else res.status(HttpStatus.BAD_REQUEST).json({ message, success });
+      if (req.user) {
+        const cartId = req.params.id as string;
+        const updateData = req.body;
+        const { message, success } = await this.cartservice.updateCart(
+          cartId,
+          updateData
+        );
+        if (success) res.status(HttpStatus.OK).json({ message, success });
+        else res.status(HttpStatus.BAD_REQUEST).json({ message, success });
+      } else
+        res.status(HttpStatus.UNAUTHORIZED).json({
+          message: responseMessage.TOKEN_ACCESS,
+          success: false,
+        });
     } catch (error) {
       res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
         response: responseMessage.ERROR_MESSAGE,
@@ -49,10 +58,16 @@ export class CartController {
 
   async deleteCart(req: AuthRequest, res: Response) {
     try {
-      const cartId = req.params.id as string;
-      const { message, success } = await this.cartservice.deleteCart(cartId);
-      if (success) res.status(HttpStatus.OK).json({ message, success });
-      else res.status(HttpStatus.BAD_REQUEST).json({ message, success });
+      if (req.user) {
+        const cartId = req.params.id as string;
+        const { message, success } = await this.cartservice.deleteCart(cartId);
+        if (success) res.status(HttpStatus.OK).json({ message, success });
+        else res.status(HttpStatus.BAD_REQUEST).json({ message, success });
+      } else
+        res.status(HttpStatus.UNAUTHORIZED).json({
+          message: responseMessage.TOKEN_ACCESS,
+          success: false,
+        });
     } catch (error) {
       res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
         response: responseMessage.ERROR_MESSAGE,
